Deck( salva mazzo-> ritoccare visualizzazione dei deck e concludere)

+++ app/src/main/java/com/example/musicdraft/data/tables/deck/Deck.kt
@@ -10,7 +10,7 @@ data class Deck (
     val nome_mazzo:String,
     val carte_associate: String,
     val email:String,
-    val popolarita: Int
+    val popolarita: Float
 
 
 ){
